refactor: inputView
